+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
     "axios": "^1.2.2",
+    "cookie": "^0.5.0",
     "crypto-js": "^4.1.1",
     "react": "^18.2.0",
     "react-calendar": "^4.0.0",
